requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0

# Аналитика и отчеты
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
reportlab==4.0.7
pillow==10.1.0
wordcloud==1.9.3

# Утилиты
python-dateutil==2.8.2
pytz==2023.3
